Removed Lint/Void with no side effect check
This has turned out to be more problematic than necessary since it
can't differentiate between actual situations (rare) and similiar
looking methods in others objects or even DSLs which should not be
effected. By removing this cop so only the defaults are used, this will
reduce some code comments that are no longer necessary.