Added check script
Necessary to check if cops -- that were previously pending but are now
enabled -- can be removed from Caliber configurations. This will help
reduce duplicated effort.

This script is being added as a development script, for now, but might
need to evolve to full fledged tooling at some point. For now, this
provides a quick fix for ensuring the Caliber configurations stay as
slim as possible.