Updated RSpec/ExampleLength to count hashes as one line
Necessary to allow arrays (default), hashes (custom), and heredocs
(default) to be counted as a single line. The downside is that this
won't prevent longer specs but reduces forcing the eye jumping up and
down the spec to find where the hash was extracted to a `let` block.
In most cases, the length of the spec remains the same. Only now all
of the information is grouped within the same spec without having to
create additional `let` blocks.