Added Metrics/CollectionLiteralLength
Necessary to limit the use of literals with massive entries which should be stored in a database instead of in memory. The default limit is 250 but that's too high so have reduced this to 25 instead. This is probably still too high still but will run with this for a while and see how it goes. I've wanted something like this for a long time so am happy to see this be part of RuboCop now. 🎉