Added Disable Syntax to check script
Necessary to ensure the custom confguration is checked against the actual configuration for changes or enhancements that might not be required to be maintained in this project.

This includes refactoring the regular expression pattern and corresponding glob since there are multiple entries which can't fit on a single line anymore.

Milestone: minor